+++ app/controllers/participants_controller.rb
@@ -118,6 +118,25 @@ class ParticipantsController < ApplicationController
     end
   end
 
+  # Returns all teammates grouped by course for the current user.
+  # GET /participants/teammates
+  def teammates
+    render json: AssignmentParticipant.all_teammates(current_user), status: :ok
+  end
+
+  # Returns the timeline of due dates and submitted review activity for a participant.
+  # GET /participants/:id/timeline
+  def timeline
+    participant = AssignmentParticipant.find_by(id: params[:id])
+    if participant.nil?
+      render json: { error: 'Participant not found' }, status: :not_found
+    elsif participant.user_id != current_user.id
+      render json: { error: 'Unauthorized' }, status: :forbidden
+    else
+      render json: participant.timeline_events, status: :ok
+    end
+  end
+
   # Permitted parameters for creating a Participant object
   def participant_params
     params.require(:participant).permit(:user_id, :assignment_id, :authorization, :can_submit,
