+++ app/models/assignment_participant.rb
@@ -50,6 +50,68 @@ class AssignmentParticipant < Participant
   end
 
   def aggregate_teammate_review_grade(teammate_review_mappings)
-    AssignmentTeam.compute_average_reviewer_score(teammate_review_mappings)
+    ResponseMap.compute_average_reviewer_score(teammate_review_mappings)
+  end
+
+  # Returns a hash of { course_name => [teammate_fullnames] } for all
+  # assignment teams the user has been part of.
+  def self.all_teammates(user)
+    result = {}
+    user.teams.each do |team|
+      next unless team.is_a?(AssignmentTeam)
+      assignment = Assignment.find_by(id: team.parent_id)
+      next if assignment.nil? || assignment[:is_calibrated]
+
+      course_name = assignment.course&.name || 'Unknown Course'
+      teammates = team.users.where.not(id: user.id).map(&:full_name)
+      next if teammates.empty?
+
+      result[course_name] ||= []
+      result[course_name] |= teammates
+    end
+    result.transform_values(&:sort).sort.to_h
+  end
+
+  # Builds a unified timeline of due dates and submitted review activity for this participant.
+  # Due dates are labeled with round info; peer reviews and author feedback appear as timestamped entries.
+  def timeline_events
+    timeline = []
+
+    DueDate.fetch_due_dates(assignment).each do |due_date|
+      timeline << {
+        'id'    => nil,
+        'name'  => due_date.round && due_date.round > 1 ? "#{due_date.deadline_name} deadline (round #{due_date.round})" : "#{due_date.deadline_name} deadline",
+        'date'  => due_date.due_at.iso8601,
+        'type'  => due_date.deadline_type_id,
+        'round' => due_date.round
+      }
+    end
+
+    ReviewResponseMap.where(reviewer_id: id).find_each do |map|
+      Response.where(map_id: map.id, is_submitted: true).order(updated_at: :desc).each do |response|
+        timeline << {
+          'id'    => response.id,
+          'name'  => "Round #{response.round} peer review",
+          'date'  => response.updated_at.iso8601,
+          'type'  => ExpertizaConstants::DeadlineTypes::REVIEW,
+          'round' => response.round
+        }
+      end
+    end
+
+    FeedbackResponseMap.where(reviewer_id: id).find_each do |map|
+      response = Response.where(map_id: map.id, is_submitted: true).order(updated_at: :desc).first
+      next if response.nil?
+
+      timeline << {
+        'id'    => response.id,
+        'name'  => 'Author feedback',
+        'date'  => response.updated_at.iso8601,
+        'type'  => nil,
+        'round' => response.round
+      }
+    end
+
+    timeline.sort_by { |item| item['date'] }
   end
 end
