+++ spec/models/student_task_spec.rb
@@ -242,199 +242,4 @@ RSpec.describe StudentTask, type: :model do
     end
   end
 
-  describe ".all_teammates" do
-    let!(:institution)     { Institution.create!(name: 'NCSU') }
-    let!(:instructor_role) { Role.find_by(name: 'Instructor') || Role.create!(name: 'Instructor') }
-    let!(:student_role)    { Role.find_by(name: 'Student')    || Role.create!(name: 'Student') }
-    let!(:instructor)      { User.create!(name: 'ts_instructor', email: 'ts_inst@test.com', full_name: 'TS Inst', password: 'password', role_id: instructor_role.id, institution: institution) }
-    let!(:course)          { Course.create!(name: 'CSC 517', directory_path: 'csc517', instructor: instructor, institution: institution) }
-    let!(:assignment)      { Assignment.create!(name: 'TS Assignment', instructor: instructor, course: course, is_calibrated: false) }
-    let!(:user)            { User.create!(name: 'ts_student1', email: 'ts_s1@test.com', full_name: 'TS Student One', password: 'password', role_id: student_role.id, institution: institution) }
-    let!(:teammate)        { User.create!(name: 'ts_student2', email: 'ts_s2@test.com', full_name: 'TS Student Two', password: 'password', role_id: student_role.id, institution: institution) }
-
-    # Helper to add a user to a team via both TeamsUser and TeamsParticipant
-    def add_to_team(team, user, assignment)
-      TeamsUser.create!(team_id: team.id, user_id: user.id)
-      participant = AssignmentParticipant.find_or_create_by!(user_id: user.id, parent_id: assignment.id) do |p|
-        p.handle = user.name
-      end
-      TeamsParticipant.find_or_create_by!(team_id: team.id, user_id: user.id, participant_id: participant.id)
-    end
-
-    it "returns teammates grouped by course name" do
-      team = AssignmentTeam.create!(name: 'Team 1', parent_id: assignment.id)
-      add_to_team(team, user, assignment)
-      add_to_team(team, teammate, assignment)
-
-      result = StudentTask.all_teammates(user)
-      expect(result).to have_key('CSC 517')
-      expect(result['CSC 517']).to include('TS Student Two')
-    end
-
-    it "excludes calibrated assignments" do
-      calibrated = Assignment.create!(name: 'Calibrated', instructor: instructor, course: course)
-      team = AssignmentTeam.create!(name: 'Team Cal', parent_id: calibrated.id)
-      add_to_team(team, user, calibrated)
-      add_to_team(team, teammate, calibrated)
-      # is_calibrated is a virtual attr_accessor — stub it on the found record
-      allow(Assignment).to receive(:find_by).with(id: calibrated.id).and_return(
-        instance_double(Assignment, is_calibrated: true, nil?: false)
-      )
-
-      result = StudentTask.all_teammates(user)
-      expect(result.values.flatten).not_to include('TS Student Two')
-    end
-
-    it "excludes the user themselves from teammates" do
-      team = AssignmentTeam.create!(name: 'Team Solo', parent_id: assignment.id)
-      add_to_team(team, user, assignment)
-
-      result = StudentTask.all_teammates(user)
-      expect(result).to be_empty
-    end
-
-    it "returns teammates sorted alphabetically" do
-      student3 = User.create!(name: 'ts_student3', email: 'ts_s3@test.com', full_name: 'Aaron Smith', password: 'password', role_id: student_role.id, institution: institution)
-      team = AssignmentTeam.create!(name: 'Team Sort', parent_id: assignment.id)
-      add_to_team(team, user, assignment)
-      add_to_team(team, teammate, assignment)
-      add_to_team(team, student3, assignment)
-
-      result = StudentTask.all_teammates(user)
-      expect(result['CSC 517']).to eq(result['CSC 517'].sort)
-    end
-  end
-
-  describe ".timeline_events" do
-    let!(:institution)  { Institution.find_by(name: 'NCSU') || Institution.create!(name: 'NCSU') }
-    let!(:inst_role)    { Role.find_by(name: 'Instructor') || Role.create!(name: 'Instructor') }
-    let!(:tl_instructor) { User.create!(name: 'tl_inst', email: 'tl_inst@test.com', full_name: 'TL Inst', password: 'password', role_id: inst_role.id, institution: institution) }
-    let!(:assignment)   { Assignment.create!(name: 'Timeline Assignment', instructor: tl_instructor) }
-    let(:participant)   { double('participant', id: 1) }
-
-    it "includes due dates with id: nil" do
-      DueDate.create!(parent: assignment, due_at: 7.days.from_now, deadline_name: 'Submission',
-                      deadline_type_id: 1, submission_allowed_id: 3, review_allowed_id: 3)
-      review_relation = double('relation')
-      allow(review_relation).to receive(:find_each)
-      allow(ReviewResponseMap).to receive(:where).with(reviewer_id: 1).and_return(review_relation)
-      feedback_relation = double('feedback_relation')
-      allow(feedback_relation).to receive(:find_each)
-      allow(FeedbackResponseMap).to receive(:where).with(reviewer_id: 1).and_return(feedback_relation)
-
-      timeline = StudentTask.timeline_events(assignment, participant)
-      submission_entry = timeline.find { |t| t['name'].include?('Submission') }
-      expect(submission_entry).not_to be_nil
-      expect(submission_entry['id']).to be_nil
-    end
-
-    it "includes submitted peer review responses with a real id" do
-      map = double('map', id: 1)
-      submitted_response = double('response', id: 99, round: 1, updated_at: 2.days.ago)
-      review_relation = double('relation')
-      allow(review_relation).to receive(:find_each).and_yield(map)
-      allow(ReviewResponseMap).to receive(:where).with(reviewer_id: 1).and_return(review_relation)
-
-      # The fix: queries for is_submitted: true with explicit ordering
-      submitted_scope = double('submitted_scope')
-      allow(submitted_scope).to receive(:each).and_yield(submitted_response)
-      ordered_scope = double('ordered_scope')
-      allow(ordered_scope).to receive(:each).and_yield(submitted_response)
-      allow(Response).to receive(:where).with(map_id: 1, is_submitted: true).and_return(ordered_scope)
-      allow(ordered_scope).to receive(:order).with(updated_at: :desc).and_return(ordered_scope)
-
-      feedback_relation = double('feedback_relation')
-      allow(feedback_relation).to receive(:find_each)
-      allow(FeedbackResponseMap).to receive(:where).with(reviewer_id: 1).and_return(feedback_relation)
-
-      timeline = StudentTask.timeline_events(assignment, participant)
-      review_entry = timeline.find { |t| t['name'].include?('peer review') }
-      expect(review_entry).not_to be_nil
-      expect(review_entry['id']).to eq(99)
-    end
-
-    it "excludes unsubmitted/draft peer review responses from the timeline" do
-      map = double('map', id: 1)
-      draft_response = double('draft_response', id: 55, round: 1, is_submitted: false, updated_at: 1.day.ago)
-
-      review_relation = double('relation')
-      allow(review_relation).to receive(:find_each).and_yield(map)
-      allow(ReviewResponseMap).to receive(:where).with(reviewer_id: 1).and_return(review_relation)
-
-      # The submitted scope returns nothing — draft response is excluded
-      empty_ordered = double('empty_ordered')
-      allow(empty_ordered).to receive(:each)
-      allow(empty_ordered).to receive(:order).with(updated_at: :desc).and_return(empty_ordered)
-      allow(Response).to receive(:where).with(map_id: 1, is_submitted: true).and_return(empty_ordered)
-
-      feedback_relation = double('feedback_relation')
-      allow(feedback_relation).to receive(:find_each)
-      allow(FeedbackResponseMap).to receive(:where).with(reviewer_id: 1).and_return(feedback_relation)
-
-      timeline = StudentTask.timeline_events(assignment, participant)
-      expect(timeline.any? { |t| t['name'].include?('peer review') }).to be false
-    end
-
-    it "captures both round 1 and round 2 responses for the same map" do
-      # The original code used Response.where(map_id:).last which silently dropped round 1
-      # when a map had two submitted responses. The fix iterates all submitted responses.
-      map = double('map', id: 1)
-      r1 = double('r1', id: 10, round: 1, updated_at: 10.days.ago)
-      r2 = double('r2', id: 11, round: 2, updated_at: 3.days.ago)
-
-      review_relation = double('relation')
-      allow(review_relation).to receive(:find_each).and_yield(map)
-      allow(ReviewResponseMap).to receive(:where).with(reviewer_id: 1).and_return(review_relation)
-
-      ordered_scope = double('ordered_scope')
-      allow(ordered_scope).to receive(:each).and_yield(r1).and_yield(r2)
-      allow(ordered_scope).to receive(:order).with(updated_at: :desc).and_return(ordered_scope)
-      allow(Response).to receive(:where).with(map_id: 1, is_submitted: true).and_return(ordered_scope)
-
-      feedback_relation = double('feedback_relation')
-      allow(feedback_relation).to receive(:find_each)
-      allow(FeedbackResponseMap).to receive(:where).with(reviewer_id: 1).and_return(feedback_relation)
-
-      timeline = StudentTask.timeline_events(assignment, participant)
-      review_entries = timeline.select { |t| t['name'].include?('peer review') }
-      expect(review_entries.map { |e| e['round'] }).to contain_exactly(1, 2)
-    end
-
-    it "excludes unsubmitted author feedback responses" do
-      map = double('map', id: 2)
-      review_relation = double('relation')
-      allow(review_relation).to receive(:find_each)
-      allow(ReviewResponseMap).to receive(:where).with(reviewer_id: 1).and_return(review_relation)
-
-      feedback_relation = double('feedback_relation')
-      allow(feedback_relation).to receive(:find_each).and_yield(map)
-      allow(FeedbackResponseMap).to receive(:where).with(reviewer_id: 1).and_return(feedback_relation)
-
-      # The fix: is_submitted: true + order(:desc).first returns nil → feedback skipped
-      ordered = double('ordered')
-      allow(ordered).to receive(:first).and_return(nil)
-      allow(Response).to receive(:where).with(map_id: 2, is_submitted: true).and_return(double(order: ordered))
-
-      timeline = StudentTask.timeline_events(assignment, participant)
-      expect(timeline.any? { |t| t['name'] == 'Author feedback' }).to be false
-    end
-
-    it "returns entries sorted by date" do
-      DueDate.create!(parent: assignment, due_at: 7.days.from_now, deadline_name: 'Submission',
-                      deadline_type_id: 1, submission_allowed_id: 3, review_allowed_id: 3)
-      DueDate.create!(parent: assignment, due_at: 14.days.from_now, deadline_name: 'Review',
-                      deadline_type_id: 2, submission_allowed_id: 3, review_allowed_id: 3)
-      review_relation = double('relation')
-      allow(review_relation).to receive(:find_each)
-      allow(ReviewResponseMap).to receive(:where).with(reviewer_id: 1).and_return(review_relation)
-      feedback_relation = double('feedback_relation')
-      allow(feedback_relation).to receive(:find_each)
-      allow(FeedbackResponseMap).to receive(:where).with(reviewer_id: 1).and_return(feedback_relation)
-
-      timeline = StudentTask.timeline_events(assignment, participant)
-      dates = timeline.map { |t| t['date'] }
-      expect(dates).to eq(dates.sort)
-    end
-  end
-
 end
