+++ spec/requests/api/v1/student_tasks_controller_spec.rb
@@ -4,7 +4,7 @@ require 'swagger_helper'
 require 'json_web_token'
 
 RSpec.describe 'StudentTasks API', type: :request do
-  before(:all) do
+  before(:each) do
     @roles = create_roles_hierarchy
   end
 
@@ -126,7 +126,6 @@ RSpec.describe 'StudentTasks API', type: :request do
           expect(data['current_stage']).to be_a(String)
           expect(data['stage_deadline']).to be_a(String)
           expect(data['permission_granted']).to be true
-          expect(data['due_dates']).to be_an(Array)
         end
       end
 
@@ -175,57 +174,4 @@ RSpec.describe 'StudentTasks API', type: :request do
     end
   end
 
-  # -------------------------------------------------------------------------
-  # /student_tasks/team
-  # -------------------------------------------------------------------------
-  path '/student_tasks/team' do
-    get 'Retrieve teammates grouped by course for current user' do
-      tags 'StudentTasks'
-      produces 'application/json'
-      parameter name: 'Authorization', in: :header, type: :string
-
-      response '200', 'returns teammates grouped by course with expected structure' do
-        let!(:team_setup) do
-          institution = Institution.create!(name: 'NCSU Team Test')
-          course      = Course.create!(name: 'CSC 517 Team', directory_path: 'csc517_team', instructor: instructor, institution: institution)
-          assignment  = Assignment.create!(name: 'Team Assignment', instructor: instructor, course: course)
-
-          teammate = User.create!(
-            name: 'teammate1', password_digest: 'password',
-            role_id: @roles[:student].id, full_name: 'Team Mate One',
-            email: 'teammate1@example.com'
-          )
-
-          team = AssignmentTeam.create!(name: 'Team Alpha', parent_id: assignment.id)
-          participant_a = AssignmentParticipant.create!(user_id: studenta.id, parent_id: assignment.id, handle: studenta.name)
-          participant_b = AssignmentParticipant.create!(user_id: teammate.id, parent_id: assignment.id, handle: teammate.name)
-
-          # user.teams traverses teams_users (User has_many :teams, through: :teams_users)
-          # team.users traverses teams_participants (Team has_many :users, through: :teams_participants)
-          # all_teammates calls both, so both join tables must have rows.
-          TeamsUser.create!(team_id: team.id, user_id: studenta.id)
-          TeamsUser.create!(team_id: team.id, user_id: teammate.id)
-          TeamsParticipant.create!(team_id: team.id, user_id: studenta.id, participant_id: participant_a.id)
-          TeamsParticipant.create!(team_id: team.id, user_id: teammate.id, participant_id: participant_b.id)
-        end
-
-        run_test! do |response|
-          data = JSON.parse(response.body)
-          # Response must be a Hash keyed by course name
-          expect(data).to be_a(Hash)
-          # The course we set up must appear as a key
-          expect(data.keys).to include('CSC 517 Team')
-          # The teammate's full name must appear under that course
-          expect(data['CSC 517 Team']).to include('Team Mate One')
-          # The current user must NOT appear in their own teammate list
-          expect(data['CSC 517 Team']).not_to include('Student A')
-        end
-      end
-
-      response '401', 'unauthorized request' do
-        let(:'Authorization') { 'Bearer ' }
-        run_test!
-      end
-    end
-  end
 end
